pyproject.toml
[tool.poetry]
name = "trade-electro-platform-project"
version = "0.1.0"
description = ""
authors = ["avdeevaa <tess-avdeeva@seznam.cz>"]
readme = "README.md"
packages = [{include = "trade_electro_platform_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.3"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
